% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDK-package.R
\docType{data}
\name{benches}
\alias{benches}
\title{Data frame of longitude-latitude points of public benches in Copenhagen}
\format{A data frame with 2124 rows and 2 variables:
\describe{
  \item{lon}{Longitude}
  \item{lat}{Latitude}
}}
\source{
Open Data Copenhagen
}
\usage{
benches
}
\description{
A dataset containing longitude-latitude points of public benches in Copenhagen
}
\keyword{datasets}

