% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapDK-package.R
\docType{data}
\name{crime}
\alias{crime}
\title{Data frame of municipal burglaries}
\format{A data frame with 98 rows and 5 variables:
\describe{
  \item{indbrud}{Number of burglaries}
  \item{year}{Year}
  \item{kommune}{Municipality name}
  \item{KOMNAVN}{Municipality name}
  \item{name}{Name}
}}
\source{
Statistics Denmark
}
\usage{
crime
}
\description{
A dataset containing number of burglaries at the municipality level from 2009
}
\keyword{datasets}

