% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getID.R
\name{getID}
\alias{getID}
\title{Print map ID}
\usage{
getID(detail = "municipal")
}
\arguments{
\item{detail}{A string specifying the detail level of the map}
}
\description{
Print id column in the \code{mapDK} data
}
\examples{
getID()
getID(detail = "polling")
}
\author{
Sebastian Barfort (\email{sebastianbarfort@gmail.com})
}
\seealso{
\url{https://github.com/sebastianbarfort/mapDK}
}

