% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapDK.R
\name{mapDK}
\alias{mapDK}
\title{Maps of Denmark}
\usage{
mapDK(values = NULL, id = NULL, data, detail = "municipal",
  show_missing = TRUE, sub = NULL, guide.label = NULL)
}
\description{
Maps of Denmark
}

