% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapDK-package.R
\docType{data}
\name{municipality}
\alias{municipality}
\title{Data frame of Denmark's 98 municipalities}
\format{A data frame with 39184 rows and 4 variables:
\describe{
  \item{long}{Longitude of municipality}
  \item{lat}{Latitude of municipality}
  \item{group}{Groups, generated by \code{ggmap:fortify}}
  \item{id}{Name of municipality}
}}
\source{
\url{Statistics Denmark}
}
\usage{
municipality
}
\description{
A dataset containing the longitude and latitude of all of Denmark's 98 municipalities
}
\keyword{datasets}

