% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapDK-package.R
\docType{data}
\name{parish}
\alias{parish}
\title{Data frame of Denmark's 1931 parishes}
\format{A data frame with 76156 rows and 4 variables:
\describe{
  \item{long}{Longitude of parish}
  \item{lat}{Latitude of parish}
  \item{group}{Groups, generated by \code{ggmap:fortify}}
  \item{id}{Identification number of parish}
}}
\source{
\url{Statistics Denmark}
}
\usage{
parish
}
\description{
A dataset containing the longitude and latitude of all of Denmark's 1931 parishes
}
\keyword{datasets}

