% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapDK-package.R
\docType{data}
\name{polling}
\alias{polling}
\title{Data frame of Denmark's 1385 polling places}
\format{A data frame with 155072 rows and 9 variables:
\describe{
  \item{long}{Longitude of polling place}
  \item{lat}{Latitude of polling place}
  \item{group}{Groups, generated by \code{ggmap:fortify}}
  \item{AfstemKod}{Longitude of polling place}
  \item{id}{Id of polling place}
  \item{KommuneNum}{Polling place municipality number}
  \item{KommuneNav}{Polling place municipality name}
  \item{OpstilNum}{...}
  \item{OpstilNav}{...}
}}
\source{
\url{Statistics Denmark}
}
\usage{
polling
}
\description{
A dataset containing the longitude and latitude of all of Denmark's 1385 polling places
}
\keyword{datasets}

