% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapDK-package.R
\docType{data}
\name{region}
\alias{region}
\title{Data frame of Denmark's 5 regions}
\format{A data frame with 32522 rows and 4 variables:
\describe{
  \item{long}{Longitude of region}
  \item{lat}{Latitude of region}
  \item{group}{Groups, generated by \code{ggmap:fortify}}
  \item{id}{Name of region}
}}
\source{
\url{Statistics Denmark}
}
\usage{
region
}
\description{
A dataset containing the longitude and latitude of all of Denmark's 5 regions
}
\keyword{datasets}

