% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapDK-package.R
\docType{data}
\name{rural}
\alias{rural}
\title{Data frame of Denmark's 11 rural areas}
\format{A data frame with 4083 rows and 4 variables:
\describe{
  \item{long}{Longitude of rural area}
  \item{lat}{Latitude of rural area}
  \item{group}{Groups, generated by \code{ggmap:fortify}}
  \item{id}{Name of rural area}
}}
\source{
\url{Statistics Denmark}
}
\usage{
rural
}
\description{
A dataset containing the longitude and latitude of all of Denmark's 11 rural areas
}
\keyword{datasets}

