% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapDK-package.R
\docType{data}
\name{votes}
\alias{votes}
\title{Data frame of Danish 2011 election results}
\format{A data frame with 13774 rows and 3 variables:
\describe{
  \item{id}{Id of polling place}
  \item{navn}{Party name}
  \item{stemmer}{Votes (in pct)}
}}
\source{
\url{http://www.dst.dk/valg/Valg1204271/xml/fintal.xml}
}
\usage{
votes
}
\description{
A dataset containing election results from the Danish general election in 2011 at the polling place level for all major parties
}
\keyword{datasets}

