% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapDK-package.R
\docType{data}
\name{zip}
\alias{zip}
\title{Data frame of Denmark's 598 zip code areas}
\format{A data frame with 49322 rows and 4 variables:
\describe{
  \item{long}{Longitude of zip code area}
  \item{lat}{Latitude of zip code area}
  \item{group}{Groups, generated by \code{ggmap:fortify}}
  \item{id}{Identification number of zip code area}
}}
\source{
\url{Statistics Denmark}
}
\usage{
zip
}
\description{
A dataset containing the longitude and latitude of all of Denmark's 598 zip oode areas
}
\keyword{datasets}

